fix custom header set (#1054)
